projects: refactor into sync_utils
